feat: change location of alert toast
